Move Connect and FallibleConnect into client module
